solving critical bugs
First, color bounds checking was wrong because of a right shift instead
of a left shift. But the more serious bug was a division by 2 (to
convert byte sizes to words) in a loop made the upper half of the MIF
code memory to not load, and go essentially zeroed out in both Code and
Data.